Move countdown tools to own directory
