+++ include/uit/countdown/CountdownTimer.hpp
@@ -4,7 +4,8 @@
 #include <chrono>
 #include <stddef.h>
 
-#include "chrono_utils.hpp"
+#include "../utility/chrono_utils.hpp"
+
 #include "CountdownWrapper.hpp"
 
 namespace uit {
