+++ tests/Makefile
@@ -3,6 +3,7 @@ SHELL := /bin/bash
 TARGET_NAMES += concurrent
 TARGET_NAMES += conduit
 TARGET_NAMES += containers
+TARGET_NAMES += countdown
 TARGET_NAMES += debug
 TARGET_NAMES += distributed
 TARGET_NAMES += mesh
